requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.5
datasets==4.0.0
dill==0.3.8
et_xmlfile==2.0.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gunicorn==21.2.0
hf-xet==1.1.5
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.1.0
networkx==3.5
numpy==1.26.4
onnx==1.18.0
onnxruntime==1.22.0
openpyxl==3.1.5
optimum==1.27.0
packaging==24.2
pandas==2.2.2
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2023.2.7
pillow==10.3.0
propcache==0.3.2
protobuf==5.27.4
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyMuPDF==1.26.3
pypdf==5.8.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rpds-py==0.27.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.37.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==4.0.2
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
